calibr/backend/requirements.txt
# ─────────────────────────────────────────────
#  Calibr – Backend Python Dependencies
#  Embeddings: Local Ollama / Groq
#  LLM:        Groq (llama3) / Ollama
#  Vector DB:  ChromaDB (local)
#  Database:   MongoDB Atlas (pymongo / motor)
# ─────────────────────────────────────────────

# --- Web framework & server ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9       # needed for file uploads (resume PDF/DOCX)

# --- Environment & config ---
python-dotenv==1.0.1

# --- LangChain core ---
langchain>=0.2.6
langchain-core>=0.2.10
langchain-community>=0.2.6

# --- LLM Providers (Groq + Ollama) ---
langchain-groq
langchain-ollama

# --- Vector store (local, no signup) ---
chromadb>=0.5.0

# --- Document parsing ---
pypdf==4.2.0                  # parse PDF resumes
python-docx==1.1.0            # parse DOCX resumes

# --- MongoDB drivers ---
pymongo==4.7.2                # sync driver for setup/migrations
motor==3.4.0                  # async driver for FastAPI routes

# --- Data validation ---
pydantic>=2.7.4
pydantic-settings>=2.3.1      # for reading .env into settings model

# --- Authentication & hashing ---
passlib[bcrypt]==1.7.4        # secure user password hashing
PyJWT==2.8.0                  # JWT generation and validation

# --- Background scheduler ---
apscheduler==3.10.4           # daily job-fetch cron at 7 AM

# --- HTTP client ---
httpx==0.27.0                 # async HTTP for Adzuna / JSearch APIs
